requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
langchain-mcp-adapters

fastapi
starlette
uvicorn
pydantic

python-dotenv
async-timeout

psycopg2-binary

numpy
pandas
scipy
matplotlib

mcp
openai
uv
dashscope
